Remove unnecessary `Mutex` protocol (that'll probably conflict with Swift 6's version anyway
