+++ Sources/Vexil/Utilities/UnfairLocks.swift
@@ -27,7 +27,7 @@ import os.lock
 /// or multiply-mapped memory, the lock implementation relies on the address of
 /// the lock value and owning process.
 ///
-struct UnfairLock<State>: Mutex {
+struct UnfairLock<State> {
 
     // MARK: - Properties
 
@@ -53,6 +53,15 @@ struct UnfairLock<State>: Mutex {
         }
     }
 
+    /// Initialise the Mutex with a lock-protected sendable `initialState`.
+    ///
+    /// - Parameter
+    ///   - initialState: An initial value to store that will be protected under the lock.
+    ///
+    init(initialState: State) where State: Sendable {
+        self.init(uncheckedState: initialState)
+    }
+
     /// Perform a closure while holding this lock.
     ///
     /// This method does not enforce sendability requirement on closure body and its return type.
@@ -68,6 +77,17 @@ struct UnfairLock<State>: Mutex {
         try mutexValue.withLockUnchecked(closure)
     }
 
+    ///  Perform a sendable closure while holding this lock.
+    ///
+    /// - Parameters:
+    ///   - closure:    A sendable closure to invoke while holding this lock.
+    /// - Returns:      The return value of `closure`.
+    /// - Throws:       Anything thrown by `closure`.
+    ///
+    func withLock<R>(_ closure: @Sendable (inout State) throws -> R) rethrows -> R where R: Sendable {
+        try withLockUnchecked(closure)
+    }
+
     /// Attempt to acquire the lock, if successful, perform a closure while holding the lock.
     ///
     /// This method does not enforce sendability requirement on closure body and its return type.
@@ -81,6 +101,18 @@ struct UnfairLock<State>: Mutex {
     ///
     func withLockIfAvailableUnchecked<R>(_ closure: (inout State) throws -> R) rethrows -> R? {
         try mutexValue.withLockIfAvailableUnchecked(closure)
+    }
+
+    ///  Attempt to acquire the lock, if successful, perform a sendable closure while
+    ///  holding the lock.
+    ///
+    /// - Parameters:
+    ///   - closure:    A sendable closure to invoke while holding this lock.
+    /// - Returns:      The return value of `closure`.
+    /// - Throws:       Anything thrown by `closure`.
+    ///
+    func withLockIfAvailable<R>(_ closure: @Sendable (inout State) throws -> R) rethrows -> R? where R: Sendable {
+        try withLockIfAvailableUnchecked(closure)
     }
 
 }
@@ -153,11 +185,6 @@ private final class LegacyUnfairLock<State>: ManagedBuffer<os_unfair_lock, State
 @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
 extension OSAllocatedUnfairLock: UnfairMutex {
     typealias UnfairState = State
-}
-
-@available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
-extension OSAllocatedUnfairLock: Mutex {
-    public typealias State = State
 }
 
 #endif
